Move Allocator unit tests to better test module
Refactoring and cleanup I made while investigating #2282.

Lots of the allocator tests where tied into the GameServerAllocation
controller tests, when there was no actual need for that to be the case.

This refactor de-couples those tests from the controller to have
better isolation in the testing, make it easier to create new tests, and
make test discovery for the Allocator system easier.

Also added a few more tests for extra coverage, and cleaned up a few
pieces here and there with the migration.
