+++ pkg/gameserverallocations/controller_test.go
@@ -18,7 +18,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"strconv"
@@ -26,18 +25,15 @@ import (
 	"time"
 
 	pb "agones.dev/agones/pkg/allocation/go"
-	"agones.dev/agones/pkg/apis"
 	agonesv1 "agones.dev/agones/pkg/apis/agones/v1"
 	allocationv1 "agones.dev/agones/pkg/apis/allocation/v1"
 	multiclusterv1 "agones.dev/agones/pkg/apis/multicluster/v1"
 	"agones.dev/agones/pkg/gameservers"
 	agtesting "agones.dev/agones/pkg/testing"
 	"agones.dev/agones/pkg/util/apiserver"
-	"agones.dev/agones/pkg/util/runtime"
 	"github.com/heptiolabs/healthcheck"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -47,7 +43,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apimachinery/pkg/watch"
 	k8stesting "k8s.io/client-go/testing"
-	"k8s.io/client-go/tools/cache"
 )
 
 const (
@@ -60,11 +55,6 @@ const (
 
 func TestControllerAllocator(t *testing.T) {
 	t.Parallel()
-
-	// TODO:(markmandel) remove once feature complete with `StateAllocationFilter` feature flag, as then the test will pass with or without the flag enabled.
-	runtime.FeatureTestMutex.Lock()
-	defer runtime.FeatureTestMutex.Unlock()
-	require.NoError(t, runtime.ParseFeatures(string(runtime.FeatureStateAllocationFilter)+"=false"))
 
 	t.Run("successful allocation", func(t *testing.T) {
 		f, gsList := defaultFixtures(4)
@@ -174,319 +164,6 @@ func TestControllerAllocator(t *testing.T) {
 	})
 }
 
-func TestControllerAllocate(t *testing.T) {
-	t.Parallel()
-
-	// TODO:(markmandel) remove once feature complete with `StateAllocationFilter` feature flag, as then the test will pass with or without the flag enabled.
-	runtime.FeatureTestMutex.Lock()
-	defer runtime.FeatureTestMutex.Unlock()
-	require.NoError(t, runtime.ParseFeatures(fmt.Sprintf("%s=false", runtime.FeatureStateAllocationFilter)))
-
-	f, gsList := defaultFixtures(4)
-	c, m := newFakeController()
-	n := metav1.Now()
-	l := map[string]string{"mode": "deathmatch"}
-	a := map[string]string{"map": "searide"}
-	fam := allocationv1.MetaPatch{Labels: l, Annotations: a}
-
-	gsList[3].ObjectMeta.DeletionTimestamp = &n
-
-	m.AgonesClient.AddReactor("list", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-		return true, &agonesv1.GameServerList{Items: gsList}, nil
-	})
-
-	updated := false
-	gsWatch := watch.NewFake()
-	m.AgonesClient.AddWatchReactor("gameservers", k8stesting.DefaultWatchReactor(gsWatch, nil))
-	m.AgonesClient.AddReactor("update", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-		ua := action.(k8stesting.UpdateAction)
-		gs := ua.GetObject().(*agonesv1.GameServer)
-
-		updated = true
-		assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-		gsWatch.Modify(gs)
-
-		return true, gs, nil
-	})
-
-	ctx, cancel := agtesting.StartInformers(m)
-	defer cancel()
-
-	if err := c.Run(ctx, 1); err != nil {
-		assert.FailNow(t, err.Error())
-	}
-	// wait for it to be up and running
-	err := wait.PollImmediate(time.Second, 10*time.Second, func() (done bool, err error) {
-		return c.allocator.allocationCache.workerqueue.RunCount() == 1, nil
-	})
-	assert.NoError(t, err)
-
-	gsa := allocationv1.GameServerAllocation{ObjectMeta: metav1.ObjectMeta{Name: "gsa-1", Namespace: defaultNs},
-		Spec: allocationv1.GameServerAllocationSpec{
-			Selectors: []allocationv1.GameServerSelector{{LabelSelector: metav1.LabelSelector{MatchLabels: map[string]string{agonesv1.FleetNameLabel: f.ObjectMeta.Name}}}},
-			MetaPatch: fam,
-		}}
-	gsa.ApplyDefaults()
-	_, ok := gsa.Validate()
-	require.True(t, ok)
-
-	gs, err := c.allocator.allocate(ctx, &gsa)
-	assert.Nil(t, err)
-	assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-	assert.True(t, updated)
-	for key, value := range fam.Labels {
-		v, ok := gs.ObjectMeta.Labels[key]
-		assert.True(t, ok)
-		assert.Equal(t, v, value)
-	}
-	for key, value := range fam.Annotations {
-		v, ok := gs.ObjectMeta.Annotations[key]
-		assert.True(t, ok)
-		assert.Equal(t, v, value)
-	}
-
-	updated = false
-	gs, err = c.allocator.allocate(ctx, &gsa)
-	assert.Nil(t, err)
-	assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-	assert.True(t, updated)
-
-	updated = false
-	gs, err = c.allocator.allocate(ctx, &gsa)
-	assert.Nil(t, err)
-	assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-	assert.True(t, updated)
-
-	updated = false
-	_, err = c.allocator.allocate(ctx, &gsa)
-	assert.NotNil(t, err)
-	assert.Equal(t, ErrNoGameServer, err)
-	assert.False(t, updated)
-}
-
-func TestControllerAllocatePriority(t *testing.T) {
-	t.Parallel()
-
-	// TODO:(markmandel) remove once feature complete with `StateAllocationFilter` feature flag, as then the test will pass with or without the flag enabled.
-	runtime.FeatureTestMutex.Lock()
-	defer runtime.FeatureTestMutex.Unlock()
-	require.NoError(t, runtime.ParseFeatures(fmt.Sprintf("%s=false", runtime.FeatureStateAllocationFilter)))
-
-	run := func(t *testing.T, name string, test func(t *testing.T, c *Controller, gas *allocationv1.GameServerAllocation)) {
-		f, gsList := defaultFixtures(4)
-		c, m := newFakeController()
-
-		gsList[0].Status.NodeName = n1
-		gsList[1].Status.NodeName = n2
-		gsList[2].Status.NodeName = n1
-		gsList[3].Status.NodeName = n1
-
-		m.AgonesClient.AddReactor("list", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			return true, &agonesv1.GameServerList{Items: gsList}, nil
-		})
-
-		gsWatch := watch.NewFake()
-		m.AgonesClient.AddWatchReactor("gameservers", k8stesting.DefaultWatchReactor(gsWatch, nil))
-		m.AgonesClient.AddReactor("update", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			ua := action.(k8stesting.UpdateAction)
-			gs := ua.GetObject().(*agonesv1.GameServer)
-			gsWatch.Modify(gs)
-
-			return true, gs, nil
-		})
-
-		ctx, cancel := agtesting.StartInformers(m)
-		defer cancel()
-
-		if err := c.Run(ctx, 1); err != nil {
-			assert.FailNow(t, err.Error())
-		}
-		// wait for it to be up and running
-		err := wait.PollImmediate(time.Second, 10*time.Second, func() (done bool, err error) {
-			return c.allocator.allocationCache.workerqueue.RunCount() == 1, nil
-		})
-		assert.NoError(t, err)
-
-		gsa := &allocationv1.GameServerAllocation{ObjectMeta: metav1.ObjectMeta{Name: "fa-1", Namespace: defaultNs},
-			Spec: allocationv1.GameServerAllocationSpec{
-				Selectors: []allocationv1.GameServerSelector{{LabelSelector: metav1.LabelSelector{MatchLabels: map[string]string{agonesv1.FleetNameLabel: f.ObjectMeta.Name}}}},
-			}}
-		gsa.ApplyDefaults()
-		_, ok := gsa.Validate()
-		require.True(t, ok)
-
-		t.Run(name, func(t *testing.T) {
-			test(t, c, gsa.DeepCopy())
-		})
-	}
-
-	run(t, "packed", func(t *testing.T, c *Controller, gas *allocationv1.GameServerAllocation) {
-		ctx := context.Background()
-		// priority should be node1, then node2
-		gs1, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.Equal(t, n1, gs1.Status.NodeName)
-
-		gs2, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.Equal(t, n1, gs2.Status.NodeName)
-		assert.NotEqual(t, gs1.ObjectMeta.Name, gs2.ObjectMeta.Name)
-
-		gs3, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.Equal(t, n1, gs3.Status.NodeName)
-		assert.NotContains(t, []string{gs1.ObjectMeta.Name, gs2.ObjectMeta.Name}, gs3.ObjectMeta.Name)
-
-		gs4, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.Equal(t, n2, gs4.Status.NodeName)
-		assert.NotContains(t, []string{gs1.ObjectMeta.Name, gs2.ObjectMeta.Name, gs3.ObjectMeta.Name}, gs4.ObjectMeta.Name)
-
-		// should have none left
-		_, err = c.allocator.allocate(ctx, gas)
-		assert.Equal(t, err, ErrNoGameServer)
-	})
-
-	run(t, "distributed", func(t *testing.T, c *Controller, gas *allocationv1.GameServerAllocation) {
-		// make a copy, to avoid the race check
-		gas = gas.DeepCopy()
-		gas.Spec.Scheduling = apis.Distributed
-
-		// distributed is randomised, so no set pattern
-		ctx := context.Background()
-
-		gs1, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-
-		gs2, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.NotEqual(t, gs1.ObjectMeta.Name, gs2.ObjectMeta.Name)
-
-		gs3, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.NotContains(t, []string{gs1.ObjectMeta.Name, gs2.ObjectMeta.Name}, gs3.ObjectMeta.Name)
-
-		gs4, err := c.allocator.allocate(ctx, gas)
-		assert.NoError(t, err)
-		assert.NotContains(t, []string{gs1.ObjectMeta.Name, gs2.ObjectMeta.Name, gs3.ObjectMeta.Name}, gs4.ObjectMeta.Name)
-
-		// should have none left
-		_, err = c.allocator.allocate(ctx, gas)
-		assert.Equal(t, err, ErrNoGameServer)
-	})
-}
-
-func TestControllerRunLocalAllocations(t *testing.T) {
-	t.Parallel()
-
-	t.Run("no problems", func(t *testing.T) {
-		f, gsList := defaultFixtures(5)
-		gsList[0].Status.NodeName = "special"
-
-		c, m := newFakeController()
-		m.AgonesClient.AddReactor("list", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			return true, &agonesv1.GameServerList{Items: gsList}, nil
-		})
-		updateCount := 0
-		m.AgonesClient.AddReactor("update", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			updateCount++
-
-			uo := action.(k8stesting.UpdateAction)
-			gs := uo.GetObject().(*agonesv1.GameServer)
-
-			return true, gs, nil
-		})
-
-		ctx, cancel := agtesting.StartInformers(m, c.allocator.allocationCache.gameServerSynced)
-		defer cancel()
-
-		// This call initializes the cache
-		err := c.allocator.allocationCache.syncCache()
-		assert.Nil(t, err)
-
-		err = c.allocator.allocationCache.counter.Run(ctx, 0)
-		assert.Nil(t, err)
-
-		gsa := &allocationv1.GameServerAllocation{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: defaultNs,
-			},
-			Spec: allocationv1.GameServerAllocationSpec{
-				Selectors: []allocationv1.GameServerSelector{{LabelSelector: metav1.LabelSelector{MatchLabels: map[string]string{agonesv1.FleetNameLabel: f.ObjectMeta.Name}}}},
-			}}
-		gsa.ApplyDefaults()
-		_, ok := gsa.Validate()
-		require.True(t, ok)
-
-		// line up 3 in a batch
-		j1 := request{gsa: gsa.DeepCopy(), response: make(chan response)}
-		c.allocator.pendingRequests <- j1
-		j2 := request{gsa: gsa.DeepCopy(), response: make(chan response)}
-		c.allocator.pendingRequests <- j2
-		j3 := request{gsa: gsa.DeepCopy(), response: make(chan response)}
-		c.allocator.pendingRequests <- j3
-
-		go c.allocator.ListenAndAllocate(ctx, 3)
-
-		res1 := <-j1.response
-		assert.NoError(t, res1.err)
-		assert.NotNil(t, res1.gs)
-
-		// since we gave gsList[0] a different nodename, it should always come first
-		assert.Contains(t, []string{"gs2", "gs3", "gs4", "gs5"}, res1.gs.ObjectMeta.Name)
-		assert.Equal(t, agonesv1.GameServerStateAllocated, res1.gs.Status.State)
-
-		res2 := <-j2.response
-		assert.NoError(t, res2.err)
-		assert.NotNil(t, res2.gs)
-		assert.NotEqual(t, res1.gs.ObjectMeta.Name, res2.gs.ObjectMeta.Name)
-		assert.Equal(t, agonesv1.GameServerStateAllocated, res2.gs.Status.State)
-
-		res3 := <-j3.response
-		assert.NoError(t, res3.err)
-		assert.NotNil(t, res3.gs)
-		assert.Equal(t, agonesv1.GameServerStateAllocated, res3.gs.Status.State)
-		assert.NotEqual(t, res1.gs.ObjectMeta.Name, res3.gs.ObjectMeta.Name)
-		assert.NotEqual(t, res2.gs.ObjectMeta.Name, res3.gs.ObjectMeta.Name)
-
-		assert.Equal(t, 3, updateCount)
-	})
-
-	t.Run("no gameservers", func(t *testing.T) {
-		c, m := newFakeController()
-		ctx, cancel := agtesting.StartInformers(m, c.allocator.allocationCache.gameServerSynced)
-		defer cancel()
-
-		// This call initializes the cache
-		err := c.allocator.allocationCache.syncCache()
-		assert.Nil(t, err)
-
-		err = c.allocator.allocationCache.counter.Run(ctx, 0)
-		assert.Nil(t, err)
-
-		gsa := &allocationv1.GameServerAllocation{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: defaultNs,
-			},
-			Spec: allocationv1.GameServerAllocationSpec{
-				Selectors: []allocationv1.GameServerSelector{{LabelSelector: metav1.LabelSelector{MatchLabels: map[string]string{agonesv1.FleetNameLabel: "thereisnofleet"}}}},
-			}}
-		gsa.ApplyDefaults()
-		_, ok := gsa.Validate()
-		require.True(t, ok)
-
-		j1 := request{gsa: gsa.DeepCopy(), response: make(chan response)}
-		c.allocator.pendingRequests <- j1
-
-		go c.allocator.ListenAndAllocate(ctx, 3)
-
-		res1 := <-j1.response
-		assert.Nil(t, res1.gs)
-		assert.Error(t, res1.err)
-		assert.Equal(t, ErrNoGameServer, res1.err)
-	})
-}
-
 func TestAllocationApiResource(t *testing.T) {
 	t.Parallel()
 
@@ -511,127 +188,6 @@ func TestAllocationApiResource(t *testing.T) {
 	if assert.Len(t, list.APIResources, 1) {
 		assert.Equal(t, "gameserverallocation", list.APIResources[0].SingularName)
 	}
-}
-
-func TestControllerAllocationUpdateWorkers(t *testing.T) {
-	t.Run("no error", func(t *testing.T) {
-		c, m := newFakeController()
-
-		updated := false
-		gs1 := &agonesv1.GameServer{
-			ObjectMeta: metav1.ObjectMeta{Name: "gs1"},
-		}
-		r := response{
-			request: request{
-				gsa:      &allocationv1.GameServerAllocation{},
-				response: make(chan response),
-			},
-			gs: gs1,
-		}
-
-		m.AgonesClient.AddReactor("update", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			updated = true
-
-			uo := action.(k8stesting.UpdateAction)
-			gs := uo.GetObject().(*agonesv1.GameServer)
-
-			assert.Equal(t, gs1.ObjectMeta.Name, gs.ObjectMeta.Name)
-			assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-
-			return true, gs, nil
-		})
-
-		updateQueue := c.allocator.allocationUpdateWorkers(context.Background(), 1)
-
-		go func() {
-			updateQueue <- r
-		}()
-
-		r = <-r.request.response
-
-		assert.True(t, updated)
-		assert.NoError(t, r.err)
-		assert.Equal(t, gs1.ObjectMeta.Name, r.gs.ObjectMeta.Name)
-		assert.Equal(t, agonesv1.GameServerStateAllocated, r.gs.Status.State)
-
-		agtesting.AssertEventContains(t, m.FakeRecorder.Events, "Allocated")
-
-		// make sure we can do more allocations than number of workers
-		gs2 := &agonesv1.GameServer{
-			ObjectMeta: metav1.ObjectMeta{Name: "gs1"},
-		}
-		r = response{
-			request: request{
-				gsa:      &allocationv1.GameServerAllocation{},
-				response: make(chan response),
-			},
-			gs: gs2,
-		}
-
-		go func() {
-			updateQueue <- r
-		}()
-
-		r = <-r.request.response
-
-		assert.True(t, updated)
-		assert.NoError(t, r.err)
-		assert.Equal(t, gs2.ObjectMeta.Name, r.gs.ObjectMeta.Name)
-		assert.Equal(t, agonesv1.GameServerStateAllocated, r.gs.Status.State)
-
-		agtesting.AssertEventContains(t, m.FakeRecorder.Events, "Allocated")
-	})
-
-	t.Run("error on update", func(t *testing.T) {
-		c, m := newFakeController()
-
-		updated := false
-		gs1 := &agonesv1.GameServer{
-			ObjectMeta: metav1.ObjectMeta{Name: "gs1"},
-		}
-		key, err := cache.MetaNamespaceKeyFunc(gs1)
-		assert.NoError(t, err)
-
-		_, ok := c.allocator.allocationCache.cache.Load(key)
-		assert.False(t, ok)
-
-		r := response{
-			request: request{
-				gsa:      &allocationv1.GameServerAllocation{},
-				response: make(chan response),
-			},
-			gs: gs1,
-		}
-
-		m.AgonesClient.AddReactor("update", "gameservers", func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-			updated = true
-
-			uo := action.(k8stesting.UpdateAction)
-			gs := uo.GetObject().(*agonesv1.GameServer)
-			assert.Equal(t, gs1.ObjectMeta.Name, gs.ObjectMeta.Name)
-			assert.Equal(t, agonesv1.GameServerStateAllocated, gs.Status.State)
-
-			return true, gs, errors.New("something went wrong")
-		})
-
-		updateQueue := c.allocator.allocationUpdateWorkers(context.Background(), 1)
-
-		go func() {
-			updateQueue <- r
-		}()
-
-		r = <-r.request.response
-
-		assert.True(t, updated)
-		assert.Error(t, r.err)
-		assert.Equal(t, gs1, r.gs)
-		agtesting.AssertNoEvent(t, m.FakeRecorder.Events)
-
-		var cached *agonesv1.GameServer
-		cached, ok = c.allocator.allocationCache.cache.Load(key)
-		assert.True(t, ok)
-		assert.Equal(t, gs1.ObjectMeta.Name, cached.ObjectMeta.Name)
-	})
 }
 
 func TestMultiClusterAllocationFromLocal(t *testing.T) {
@@ -1204,111 +760,6 @@ func TestMultiClusterAllocationFromRemote(t *testing.T) {
 		_, err = executeAllocation(gsa, c)
 		assert.NoError(t, err)
 		assert.Equal(t, 2, calls)
-	})
-}
-
-func TestCreateRestClientError(t *testing.T) {
-	t.Parallel()
-	t.Run("Missing secret", func(t *testing.T) {
-		c, _ := newFakeController()
-
-		connectionInfo := &multiclusterv1.ClusterConnectionInfo{
-			SecretName: "secret-name",
-		}
-		_, err := c.allocator.createRemoteClusterDialOption(defaultNs, connectionInfo)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "secret-name")
-	})
-	t.Run("Missing cert", func(t *testing.T) {
-		c, m := newFakeController()
-
-		m.KubeClient.AddReactor("list", "secrets",
-			func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-				return true, &corev1.SecretList{
-					Items: []corev1.Secret{{
-						Data: map[string][]byte{
-							"tls.crt": clientCert,
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "secret-name",
-							Namespace: defaultNs,
-						},
-					}}}, nil
-			})
-
-		_, cancel := agtesting.StartInformers(m, c.allocator.secretSynced)
-		defer cancel()
-
-		connectionInfo := &multiclusterv1.ClusterConnectionInfo{
-			SecretName: "secret-name",
-		}
-		_, err := c.allocator.createRemoteClusterDialOption(defaultNs, connectionInfo)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "missing client certificate key pair in secret secret-name")
-	})
-	t.Run("Bad client cert", func(t *testing.T) {
-		c, m := newFakeController()
-
-		m.KubeClient.AddReactor("list", "secrets",
-			func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-				return true, &corev1.SecretList{
-					Items: []corev1.Secret{{
-						Data: map[string][]byte{
-							"tls.crt": []byte("XXX"),
-							"tls.key": []byte("XXX"),
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "secret-name",
-							Namespace: defaultNs,
-						},
-					}}}, nil
-			})
-
-		_, cancel := agtesting.StartInformers(m, c.allocator.secretSynced)
-		defer cancel()
-
-		connectionInfo := &multiclusterv1.ClusterConnectionInfo{
-			SecretName: "secret-name",
-		}
-		_, err := c.allocator.createRemoteClusterDialOption(defaultNs, connectionInfo)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to find any PEM data in certificate input")
-	})
-	t.Run("Bad CA cert", func(t *testing.T) {
-		c, m := newFakeController()
-
-		m.KubeClient.AddReactor("list", "secrets",
-			func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-				return true, getTestSecret("secret-name", clientCert), nil
-			})
-
-		_, cancel := agtesting.StartInformers(m, c.allocator.secretSynced)
-		defer cancel()
-
-		connectionInfo := &multiclusterv1.ClusterConnectionInfo{
-			SecretName: "secret-name",
-			ServerCA:   []byte("XXX"),
-		}
-		_, err := c.allocator.createRemoteClusterDialOption(defaultNs, connectionInfo)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "PEM format")
-	})
-	t.Run("Bad client CA cert", func(t *testing.T) {
-		c, m := newFakeController()
-
-		m.KubeClient.AddReactor("list", "secrets",
-			func(action k8stesting.Action) (bool, k8sruntime.Object, error) {
-				return true, getTestSecret("secret-name", []byte("XXX")), nil
-			})
-
-		_, cancel := agtesting.StartInformers(m, c.allocator.secretSynced)
-		defer cancel()
-
-		connectionInfo := &multiclusterv1.ClusterConnectionInfo{
-			SecretName: "secret-name",
-		}
-		_, err := c.allocator.createRemoteClusterDialOption(defaultNs, connectionInfo)
-		assert.Nil(t, err)
 	})
 }
